Fix container ignore files
