+++ pkg/logs/internal/launchers/file/launcher.go
@@ -6,10 +6,7 @@
 package file
 
 import (
-	"os"
-	"path/filepath"
 	"regexp"
-	"strings"
 	"time"
 
 	"github.com/DataDog/datadog-agent/pkg/util"
@@ -26,17 +23,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/logs/sources"
 	"github.com/DataDog/datadog-agent/pkg/util/startstop"
 )
-
-// rxContainerID is used in the shouldIgnore func to do a best-effort validation
-// that the file currently scanned for a source is attached to the proper container.
-// If the container ID we parse from the filename isn't matching this regexp, we *will*
-// tail the file because we prefer a false-negative than a false-positive (best-effort).
-var rxContainerID = regexp.MustCompile("^[a-fA-F0-9]{64}$")
-
-// ContainersLogsDir is the directory in which we should find containers logsfile
-// with the container ID in their filename.
-// Public to be able to change it while running unit tests.
-var ContainersLogsDir = "/var/log/containers"
 
 // DefaultSleepDuration represents the amount of time the tailer waits before reading new data when no data is received
 const DefaultSleepDuration = 1 * time.Second
@@ -136,7 +122,7 @@ func (s *Launcher) cleanup() {
 // For instance, when a file is logrotated, its tailer will keep tailing the rotated file.
 // The Scanner needs to stop that previous tailer, and start a new one for the new file.
 func (s *Launcher) scan() {
-	files := s.fileProvider.FilesToTail(s.activeSources)
+	files := s.fileProvider.FilesToTail(s.validatePodContainerID, s.activeSources)
 	filesTailed := make(map[string]bool)
 
 	log.Debugf("Scan - got %d files from FilesToTail and currently tailing %d files\n", len(files), len(s.tailers))
@@ -280,23 +266,6 @@ func (s *Launcher) startNewTailer(file *tailer.File, m config.TailingMode) bool 
 		return false
 	}
 
-	// We also use the file launcher to look for containers and pods logs file, because of that
-	// we have to make sure that the file we just detected is tagged with the correct
-	// container ID. Enabled through `logs_config.validate_pod_container_id`.
-	// The way k8s is storing files in /var/log/pods doesn't let us do that properly
-	// (the filename doesn't contain the container ID).
-	// However, the symlinks present in /var/log/containers are pointing to /var/log/pods files,
-	// meaning that we can use them to validate that the file we have found is concerning us.
-	// That's what the function shouldIgnore is trying to do when the directory exists and is readable.
-	// See these links for more info:
-	//   - https://github.com/kubernetes/kubernetes/issues/58638
-	//   - https://github.com/fabric8io/fluent-plugin-kubernetes_metadata_filter/issues/105
-	if s.validatePodContainerID && file.Source != nil &&
-		(file.Source.GetSourceType() == sources.KubernetesSourceType || file.Source.GetSourceType() == sources.DockerSourceType) &&
-		s.shouldIgnore(file) {
-		return false
-	}
-
 	tailer := s.createTailer(file, s.pipelineProvider.NextPipelineChan())
 
 	var offset int64
@@ -317,69 +286,6 @@ func (s *Launcher) startNewTailer(file *tailer.File, m config.TailingMode) bool 
 
 	s.tailers[file.GetScanKey()] = tailer
 	return true
-}
-
-// shouldIgnore resolves symlinks in /var/log/containers in order to use that redirection
-// to validate that we will be reading a file for the correct container.
-func (s *Launcher) shouldIgnore(file *tailer.File) bool {
-	// this method needs a source config to detect whether we should ignore that file or not
-	if file == nil || file.Source == nil || file.Source.Config() == nil {
-		return false
-	}
-
-	infos := make(map[string]string)
-	err := filepath.Walk(ContainersLogsDir, func(containerLogFilename string, info os.FileInfo, err error) error {
-		// we only wants to follow symlinks
-		if info == nil || info.Mode()&os.ModeSymlink != os.ModeSymlink || info.IsDir() {
-			// not a symlink, we are not interested in this file
-			return nil
-		}
-
-		// resolve the symlink
-		podLogFilename, err2 := os.Readlink(containerLogFilename)
-		if err2 != nil {
-			log.Debug("Error while resolving symlink of", containerLogFilename, ":", err)
-			return nil
-		}
-
-		infos[podLogFilename] = containerLogFilename
-		return nil
-	})
-
-	// this is not an error if we we are not currently looking for container logs files,
-	// so not problem and just return false.
-	// Still, we write a debug message to be able to troubleshoot that
-	// in cases we're legitimately looking for containers logs.
-	if err != nil {
-		log.Debug("Can't look for symlinks in /var/log/containers:", err)
-		return false
-	}
-
-	// container id extracted from the file found in /var/log/containers
-	base := filepath.Base(infos[file.Path]) // only the file
-	ext := filepath.Ext(base)               // file extension
-	parts := strings.Split(base, "-")       // get only the container ID part from the file
-	var containerIDFromFilename string
-	if len(parts) > 1 {
-		containerIDFromFilename = strings.TrimSuffix(parts[len(parts)-1], ext)
-	}
-
-	// basic validation of the ID that has been parsed, if it doesn't look like
-	// an ID we don't want to compare another ID to it
-	if containerIDFromFilename == "" || !rxContainerID.Match([]byte(containerIDFromFilename)) {
-		return false
-	}
-
-	if file.Source.Config().Identifier != "" && containerIDFromFilename != "" {
-		if strings.TrimSpace(strings.ToLower(containerIDFromFilename)) != strings.TrimSpace(strings.ToLower(file.Source.Config().Identifier)) {
-			log.Debugf("We were about to tail a file attached to the wrong container (%s != %s), probably due to short-lived containers.",
-				containerIDFromFilename, file.Source.Config().Identifier)
-			// ignore this file, it is not concerning the container stored in file.Source
-			return true
-		}
-	}
-
-	return false
 }
 
 // handleTailingModeChange determines the tailing behaviour when the tailing mode for a given file has its
